pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "docupilot"
version = "0.1.0"
description = "Multi-agent document OCR + normalization + risk analysis pipeline"
requires-python = ">=3.9"
dependencies = [
  "python-dotenv>=1.0.0",
  "pydantic>=2.7.0",
  "requests>=2.32.0",
  "paddlepaddle>=2.5.0",
  "paddleocr>=2.7.0",
  "pdf2image>=1.17.0",
  "Pillow>=10.0.0",
]

[project.scripts]
docupilot = "docupilot.main:cli"
